extract logic into appropriate Routes
